Profile status editing

+++ Controllers/MessagesController.cs
@@ -1,4 +1,6 @@
-﻿using DAW_social_platform.Models;
+﻿using DAW_social_platform.Infrastructure;
+using DAW_social_platform.Models;
+using Microsoft.AspNet.Identity;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -7,30 +9,42 @@ using System.Web.Mvc;
 
 namespace DAW_social_platform.Controllers
 {
+    [Authorize(Roles = "User,Admin")]
     public class MessagesController : Controller
     {
         private ApplicationDbContext db = new ApplicationDbContext();
-
-        // GET: Messages
-        public ActionResult Index()
-        {
-            return View();
-        }
+        private GroupAuthorization GroupAuth = new GroupAuthorization();
+ 
 
         [HttpDelete]
         public ActionResult Delete(int id)
         {
             Message mes = db.Messages.Find(id);
-            db.Messages.Remove(mes);
-            db.SaveChanges();
+            if (GroupAuth.IsAllowedToDelete(mes.GroupId, User.Identity.GetUserId()) || User.Identity.GetUserId() == mes.UserId || User.IsInRole("Admin"))
+            {
+                db.Messages.Remove(mes);
+                db.SaveChanges();
+            }
+            else
+            {
+                TempData["message"] = "Nu aveti dreptul sa stergeti un mesaj care nu va apartine";
+            }
             return Redirect("/Groups/Show/" + mes.GroupId);
         }
 
         public ActionResult Edit(int id)
         {
             Message mes = db.Messages.Find(id);
-            ViewBag.message = mes;
-            return View();
+            if (GroupAuth.IsAllowedToEdit(mes.GroupId, User.Identity.GetUserId()) || User.Identity.GetUserId() == mes.UserId || User.IsInRole("Admin"))
+            {
+                ViewBag.message = mes;
+                return View();
+            } 
+            else
+            {
+                TempData["message"] = "Nu aveti dreptul sa editati un mesaj care nu va apartine";
+                return Redirect("/Groups/Show/" + mes.GroupId);
+            }
         }
 
         [HttpPut]
@@ -39,10 +53,17 @@ namespace DAW_social_platform.Controllers
             try
             {
                 Message mes = db.Messages.Find(id);
-                if (TryUpdateModel(mes))
+                if (GroupAuth.IsAllowedToEdit(mes.GroupId, User.Identity.GetUserId()) || User.Identity.GetUserId() == mes.UserId || User.IsInRole("Admin"))
                 {
-                    mes.MessageContent = requestMessage.MessageContent;
-                    db.SaveChanges();
+                    if (TryUpdateModel(mes))
+                    {
+                        mes.MessageContent = requestMessage.MessageContent;
+                        db.SaveChanges();
+                    }
+                }
+                else
+                {
+                    TempData["message"] = "Nu aveti dreptul sa editati un mesaj care nu va apartine";
                 }
                 return Redirect("/Groups/Show/" + mes.GroupId);
             }
